sentiment_analyzer/requirements.txt
kafka-python
vaderSentiment
